Item data structs + Parent class + Basic item pickup and throw + gamepad input
Created foundation of the game - picking up and throwing items.

Items now vanish and become 0 scale instead of physically attaching the items (old version)

Much better UX this way

Pickup is no longer directional as im using a box trace not a line trace in front of the char.

shared input flip flop for 'Pickup' and 'Use' to simplify experience